FIX all broken tests
